test get llm provider
